requirements.txt
numpy
tensorflow
opencv-python-headless
transformers
pillow
scikit-learn
pytorch
torchvision